Update conftest to contain the mock collections to be used in other places.
